Convert deployment guide to INSTALL.md.
Convert "Outreach App Deployment Guide.docx" to markdown format and
name it INSTALL.md.  Add some notes based on an initial run-through of
the document, and add TBDs about including a build file,
instructions to create the outreach.war file that is listed as a
requirement, and documentation on deploying with Eclipse, which is the
preferred method of the founding development team.